requirements.txt
flask
google-generativeai
google-genai